requirements.txt
langchain
chromadb
streamlit
python-dotenv
openai